moved RunnerConfiguration
